refactor: long을 Long으로 변경하여 일관성 유지

+++ src/main/java/roomescape/web/api/ReservationTimeController.java
@@ -42,7 +42,7 @@ public class ReservationTimeController {
     @GetMapping("/times/available")
     public ResponseEntity<List<AvailableReservationTimeResponse>> findAllAvailable(
             @RequestParam("date") LocalDate date,
-            @RequestParam("themeId") Integer themeId
+            @RequestParam("themeId") Long themeId
     ) {
         List<AvailableReservationTimeResponse> reservationTimeResponses = reservationTimeService.findAllWithAvailability(
                 date, themeId);
